move parser to separate file and add simple eslint config
