+++ lib/vttcue-extended.js
@@ -40,13 +40,16 @@ if (typeof module !== "undefined" && module.exports) {
   };
 
   root.VTTCue.create = function(options) {
-    if (!options.hasOwnProperty("startTime") || !options.hasOwnProperty("endTime") ||
-        !options.hasOwnProperty("text")) {
+    if (
+      !Object.prototype.hasOwnProperty.call(options, "startTime") ||
+      !Object.prototype.hasOwnProperty.call(options, "endTime") ||
+      !Object.prototype.hasOwnProperty.call(options, "text")
+    ) {
       throw new Error("You must at least have start time, end time, and text.");
     }
     var cue = new root.VTTCue(options.startTime, options.endTime, options.text);
     for (var key in options) {
-      if (cue.hasOwnProperty(key)) {
+      if (Object.prototype.hasOwnProperty.call(cue, key)) {
         cue[key] = options[key];
       }
     }
