+++ lib/vttregion-extended.js
@@ -29,7 +29,7 @@ if (typeof module !== "undefined" && module.exports) {
   root.VTTRegion.create = function(options) {
     var region = new root.VTTRegion();
     for (var key in options) {
-      if (region.hasOwnProperty(key)) {
+      if (Object.prototype.hasOwnProperty.call(region, key)) {
         region[key] = options[key];
       }
     }
