+++ package.json
@@ -18,6 +18,7 @@
     "deumdify": "^1.2.4",
     "difflet": "git://github.com/humphd/difflet.git#714b82706ad39d75275a886aeff637df5097626f",
     "dive": "0.3.1",
+    "eslint": "^6.7.1",
     "grunt": "0.4.1",
     "grunt-bump": "0.0.14",
     "grunt-cli": "0.1.9",
